requirements.txt
playwright==1.40.0
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9
lxml==4.9.3
beautifulsoup4==4.12.2
requests==2.31.0
networkx
openai
apscheduler
google-api-python-client
markdown
pdfkit
flask
pillow
